api/requirements.txt
# cnc-mayyanks-api — Python Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
alembic>=1.13.0

# Auth & Security
pyjwt>=2.8.0
bcrypt>=4.1.0
cryptography>=41.0.0
httpx>=0.26.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Streaming
paho-mqtt>=1.6.1
kafka-python>=2.0.2

# Cache
redis>=5.0.0

# Observability
prometheus-client>=0.19.0
loguru>=0.7.0

# Serialization
orjson>=3.9.0

# Data processing
numpy>=1.26.0
pandas>=2.1.0
